% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_tools.R
\name{sqchangeParam}
\alias{sqchangeParam}
\title{Edit .sqvar or .sqpar XML files}
\usage{
sqchangeParam(file, varspecie, parameter, value)
}
\arguments{
\item{file}{A valid .sqvar or .sqpar Sirius Quality file.}

\item{varspecie}{The name of the target specie/variety.}

\item{parameter}{The name of the parameter to change.}

\item{value}{Value to set the parameter.}
}
\value{
The name of the edited file.
}
\description{
This function change the specified parameter in a XML .sqvar or .sqpar file
for a specific variety/specie
}
