% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_tools.R
\name{sqduplVariety}
\alias{sqduplVariety}
\title{Duplicate an existing variety/specie in XML files}
\usage{
sqduplVariety(file, varspecie, new)
}
\arguments{
\item{file}{Target .sqpar or .sqvar file}

\item{varspecie}{Variety/Specie to duplicate}

\item{new}{Name of the new Variety/Specie}
}
\description{
This function duplicates an existing variety/specie with a new name in the supplied .sqvar/.sqpar file
}
