% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_tools.R
\name{sqeditParam}
\alias{sqeditParam}
\title{Edit .sqvar or .sqpar XML files}
\usage{
sqeditParam(node, varspecie = NULL, parameter, value)
}
\arguments{
\item{node}{Object of class "XMLInternalDocument" or "XMLAbstractDocument" obtained from sqpar or sqvar file}

\item{varspecie}{The name of the target specie/variety}

\item{parameter}{The name of the parameter to change}

\item{value}{Value to set the parameter}
}
\value{
Changed XML tree
}
\description{
This function returns a XML .sqvar or .sqpar file
with the specified parameter changed for a variety/specie
with the provided value
}
