% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_tools.R
\name{sqgetObsItems}
\alias{sqgetObsItems}
\title{Get all the Observations Items in the observation file}
\usage{
sqgetObsItems(sqpro)
}
\arguments{
\item{sqpro}{Target Project file (.sqpro)}
}
\value{
A character vector containing all the present Observations Items
}
\description{
Function to retrieve all the Observations Items present in the Observation File
}
