% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_tools.R
\name{sqgetRunItems}
\alias{sqgetRunItems}
\title{Get all the run items in Run File}
\usage{
sqgetRunItems(sqrun)
}
\arguments{
\item{sqrun}{Target .sqrun file}
}
\value{
A character vector containing all the present Run Items
}
\description{
Funciton to retrieve all the Run Items present in the specified Sirius Quality Run file (.sqrun)
}
