% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtools.R
\name{sqOptimiseVar}
\alias{sqOptimiseVar}
\title{Optimise one genotype using the specified project/management(s) against one or more observations.}
\usage{
sqOptimiseVar(sqconsole, project, runitem, mng, var, parameters, ..., obsitem,
  observations)
}
\arguments{
\item{sqconsole}{Path to SiriusQuality-Console.exe.}

\item{project}{Path to Project file (can be relative).}

\item{runitem}{Character string identifying the chosen RunItem.}

\item{mng}{Character vector including all the chosen Management(s).}

\item{var}{Character string identifying the chosen Variety.}

\item{parameters}{Character string indicating the parameters to be optimized.}

\item{obsitem}{Character string identifying the chosen Observations Item.}

\item{observation}{Character vector including all the chosen Observation(s) to optimise against.}
}
\value{
A data frame containing all the optimised combinations with value(s) of the fitness function(s)
}
\description{
Optimise one genotype using the specified project/management(s) against one or more observations.
}
